An asset-based company dataset for demonstration

Fake data about physical assets (e.g. wind turbine power plant capacities), aggregated to company-level. These data are used to assess the climate alignment of financial portfolios. It imitates data from market-intelligence databases.

Demo datasets are synthetic because most financial data is strictly private; they help to demonstrate and test the implementation in R of 'PACTA' (https://2degrees-investing.org/resource/pacta/).

Usage

abcd_demo

Format

An object of class tbl_df (inherits from tbl, data.frame) with 17668 rows and 14 columns.

Definitions

abcd_timestamp (character): Date at which asset data was sourced from the data provider.

company_id (character): The id of the company owning the asset created by the data provider.

country_of_domicile (character): Country where company is registered.

emission_factor (double): Company level emission factor of the technology.

emission_factor_unit (character): The units that the emission factor is measured in.

is_ultimate_owner (logical): Flag if company is the ultimate parent in our database.

lei (character): The legal entity identifier of the company owning the asset.

name_company (character): The name of the company owning the asset.

plant_location (character): Country where asset is located.

production (double): Company level production of the technology.

production_unit (character): The units that production is measured in.

sector (character): Sector to which the asset belongs.

technology (character): Technology implemented by the asset.

year (integer): Year at which the production value is predicted.

See also

Examples

head(abcd_demo)
#> # A tibble: 6 × 14
#>   company_id name_company          lei   sector technology production_unit  year
#>   <chr>      <chr>                 <chr> <chr>  <chr>      <chr>           <int>
#> 1 1          aba hydropower gener… 8360… power  hydrocap   MW               2013
#> 2 1          aba hydropower gener… 8360… power  hydrocap   MW               2014
#> 3 1          aba hydropower gener… 8360… power  hydrocap   MW               2015
#> 4 1          aba hydropower gener… 8360… power  hydrocap   MW               2016
#> 5 1          aba hydropower gener… 8360… power  hydrocap   MW               2017
#> 6 1          aba hydropower gener… 8360… power  hydrocap   MW               2018
#> # … with 7 more variables: production <dbl>, emission_factor <dbl>,
#> #   country_of_domicile <chr>, plant_location <chr>, is_ultimate_owner <lgl>,
#> #   abcd_timestamp <chr>, emission_factor_unit <chr>