An asset level dataset for demonstration

Fake data about physical assets (e.g. wind turbine power plant capacities) used to assess the climate alignment of financial portfolios. It imitates data from market-intelligence databases.

Demo datasets are synthetic because most financial data is strictly private; they help to demonstrate and test the implementation in R of 'PACTA' (https://2degrees-investing.org/resource/pacta/).

ald_demo

Format

ald_demo is a data.frame with columns:

ald_timestamp (character): Date at which asset data was pulled from database.

country_of_domicile (character): Country where company is registered.

emission_factor (double): Company level emission factor of the technology.

is_ultimate_listed_owner (logical): Flag if company is the listed ultimate parent.

is_ultimate_owner (logical): Flag if company is the ultimate parent in our database.

name_company (character): The name of the company owning the asset.

number_of_assets (integer): Number of assets of a given technology owned by the company.

plant_location (character): Country where asset is located.

production (double): Company level production of the technology.

production_unit (character): The units that production is measured in.

sector (character): Sector to which the asset belongs.

technology (character): Technology implemented by the asset.

year (integer): Year at which the production value is predicted.

See also

Examples

head(ald_demo)

#> # A tibble: 6 x 13
#>   name_company sector technology production_unit  year production
#>   <chr>        <chr>  <chr>      <chr>           <dbl>      <dbl>
#> 1 aba hydropo… power  hydrocap   MW               2013    133340.
#> 2 aba hydropo… power  hydrocap   MW               2014    131582.
#> 3 aba hydropo… power  hydrocap   MW               2015    129824.
#> 4 aba hydropo… power  hydrocap   MW               2016    128065.
#> 5 aba hydropo… power  hydrocap   MW               2017    126307.
#> 6 aba hydropo… power  hydrocap   MW               2018    124549.
#> # … with 7 more variables: emission_factor <dbl>, country_of_domicile <chr>,
#> #   plant_location <chr>, number_of_assets <dbl>, is_ultimate_owner <lgl>,
#> #   is_ultimate_listed_owner <lgl>, ald_timestamp <chr>