Datasets to bridge (translate) common sector-classification codes

These datasets serve as a translation key between common sector-classification systems and sectors relevant to the 'PACTA' tool (https://2degrees-investing.org/resource/pacta/).

Classification datasets help to standardize sector classification codes from the wild to a relevant subset including 'power', 'oil and gas', 'coal', 'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.

isic_classification

nace_classification

naics_classification

Format

classification_bridge is a data.frame with columns:

borderline (logical): Flag indicating if 2dii sector and classification code are a borderline match.

code (character): Original ISIC code.

code_level (double): Level of granularity of ISIC code.

sector (character): Associated 2dii sector.

nace_classification is a data.frame with columns:

borderline (logical): Flag indicating if 2dii sector and classification code are a borderline match.

code (double): Formatted NACE code removing periods.

code_level (double): Level of granularity of NACE code.

original_code (double): Original NACE code.

sector (character): Associated 2dii sector.

naics_classification is a data.frame with columns:

borderline (logical): Flag indicating if 2dii sector and classification code are a borderline match.

code (double): Formatted NAICS code.

code_level (double): Level of granularity of NAICS code.

original_code (double): Original NAICS code.

sector (character): Associated 2dii sector.

See also

Examples

head(isic_classification)

#> # A tibble: 6 x 4
#>   code  code_level sector       borderline
#>   <chr>      <dbl> <chr>        <lgl>     
#> 1 A              1 not in scope FALSE     
#> 2 1              2 not in scope FALSE     
#> 3 11             3 not in scope FALSE     
#> 4 111            4 not in scope FALSE     
#> 5 112            4 not in scope FALSE     
#> 6 113            4 not in scope FALSE     

head(nace_classification)

#> # A tibble: 6 x 5
#>   original_code  code code_level sector   borderline
#>           <dbl> <dbl>      <dbl> <chr>    <lgl>     
#> 1          30.1   301          3 shipping TRUE      
#> 2          30.1  3011          4 shipping TRUE      
#> 3          30.1  3012          4 shipping TRUE      
#> 4          33.2  3315          4 shipping TRUE      
#> 5          50      50          2 shipping TRUE      
#> 6          50.1   501          4 shipping TRUE      

head(naics_classification)

#> # A tibble: 6 x 5
#>   original_code  code code_level sector   borderline
#>           <dbl> <dbl>      <dbl> <chr>    <lgl>     
#> 1          30.1   301          3 shipping TRUE      
#> 2          30.1  3011          4 shipping TRUE      
#> 3          30.1  3012          4 shipping TRUE      
#> 4          33.2  3315          4 shipping TRUE      
#> 5          50      50          2 shipping TRUE      
#> 6          50.1   501          4 shipping TRUE