A prepared co2 intensity climate scenario dataset for demonstration

Fake co2 intensity climate scenario dataset, prepared for the software PACTA (Paris Agreement Capital Transition Assessment). It imitates climate scenario data (e.g. from the International Energy Agency (IEA)) including the change through time in production across industrial sectors (calculated by 2DII).

Demo datasets are synthetic because most financial data is strictly private; they help to demonstrate and test the implementation in R of 'PACTA' (https://2degrees-investing.org/resource/pacta/).

co2_intensity_scenario_demo

Format

An object of class tbl_df (inherits from tbl, data.frame) with 22 rows and 7 columns.

Definitions

ald_emission_factor_unit (character): The units that the emission factor is measured in.

ald_timestamp (character): Date at which asset data was pulled from database.

country_of_domicile (character): Country where company is registered.

emission_factor (double): Company level emission factor of the technology.

is_ultimate_listed_owner (logical): Flag if company is the listed ultimate parent.

is_ultimate_owner (logical): Flag if company is the ultimate parent in our database.

name_company (character): The name of the company owning the asset.

number_of_assets (integer): Number of assets of a given technology owned by the company.

plant_location (character): Country where asset is located.

production (double): Company level production of the technology.

production_unit (character): The units that production is measured in.

sector (character): Sector to which the asset belongs.

technology (character): Technology implemented by the asset.

year (integer): Year at which the production value is predicted.

See also

Examples

head(co2_intensity_scenario_demo)

#> # A tibble: 6 x 7
#>   scenario_source scenario sector region  year emission_factor_… emission_factor
#>   <chr>           <chr>    <chr>  <chr>  <dbl> <chr>                       <dbl>
#> 1 demo_2020       demo     cement global  2020 tons of CO2 per …           0.7  
#> 2 demo_2020       demo     cement global  2021 tons of CO2 per …           0.64 
#> 3 demo_2020       demo     cement global  2022 tons of CO2 per …           0.580
#> 4 demo_2020       demo     cement global  2023 tons of CO2 per …           0.52 
#> 5 demo_2020       demo     cement global  2024 tons of CO2 per …           0.46 
#> 6 demo_2020       demo     cement global  2025 tons of CO2 per …           0.4