Column definitions of all datasets

This dataset provides metadata about all datasets in the package r2dii.data.

data_dictionary

Format

An object of class tbl_df (inherits from tbl, data.frame) with 104 rows and 4 columns.

Definitions

column (character): The name of a dataset-column.

dataset (character): The name of a dataset.

definition (character): The definition of a dataset-column.

typeof (character): The result of typeof(), one of double, integer, logical, or character..

Examples

head(data_dictionary)

#> # A tibble: 6 × 4
#>   dataset  column               typeof    definition                            
#>   <chr>    <chr>                <chr>     <chr>                                 
#> 1 ald_demo ald_timestamp        character Date at which asset data was sourced …
#> 2 ald_demo company_id           character The id of the company owning the asse…
#> 3 ald_demo country_of_domicile  character Country where company is registered   
#> 4 ald_demo emission_factor      double    Company level emission factor of the …
#> 5 ald_demo emission_factor_unit character The units that the emission factor is…
#> 6 ald_demo is_ultimate_owner    logical   Flag if company is the ultimate paren…