Dataset to bridge (translate) common sector-classification codes

These dataset serve as a translation key between common sector-classification systems and sectors relevant to the 'PACTA' tool (https://2degrees-investing.org/resource/pacta/).

gics_classification

Format

An object of class tbl_df (inherits from tbl, data.frame) with 263 rows and 5 columns.

Definitions

borderline (logical): Flag indicating if 2dii sector and classification code are a borderline match.

code (double): Original GICS code.

code_level (double): Level of granularity of GICS code.

description (character): Original GICS description.

sector (character): Associated 2dii sector.

Details

Classification datasets help to standardize sector classification codes from the wild to a relevant subset including 'power', 'oil and gas', 'coal', 'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.

See also

Examples

head(gics_classification)

#> # A tibble: 6 x 5
#>       code code_level sector      borderline description                       
#>      <dbl>      <dbl> <chr>       <lgl>      <chr>                             
#> 1 10101010          4 oil and gas FALSE      oil & gas drilling                
#> 2 10101020          4 oil and gas TRUE       oil & gas equipment & services    
#> 3 10102010          4 oil and gas FALSE      integrated oil & gas              
#> 4 10102020          4 oil and gas FALSE      oil & gas exploration & production
#> 5 10102030          4 oil and gas TRUE       oil & gas refining & marketing    
#> 6 10102040          4 oil and gas TRUE       oil & gas storage & transportation