Determine if a technology is green or brown

This dataset provides a simple lookup table to determine if a technology is meant to increase (green) or decrease (brown) to align with a future global warming of less than 2 degrees.

green_or_brown

Format

An object of class tbl_df (inherits from tbl, data.frame) with 16 rows and 3 columns.

Definitions

green_or_brown (character): If the technology is green (increasing) or brown (decreasing), as defined by the Paris-aligned IEA scenarios.

sector (character): The sector to which the technology belongs.

technology (character): The technology sub-category within the sector.

See also

Examples

head(green_or_brown)

#> # A tibble: 6 x 3
#>   sector     technology    green_or_brown
#>   <chr>      <chr>         <chr>         
#> 1 automotive electric      green         
#> 2 automotive hybrid        green         
#> 3 automotive ice           brown         
#> 4 automotive fuelcell      green         
#> 5 power      hydrocap      green         
#> 6 power      renewablescap green