A loanbook dataset for demonstration

This is a public dataset that may be used as a template for banks to create their own loanbook datasets. It is also useful for examples and tests.

loanbook_demo

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 320 rows and 19 columns.

See also

Examples

head(loanbook_demo)

#> # A tibble: 6 x 19
#>   id_loan id_direct_loant… name_direct_loa… id_intermediate… name_intermedia…
#>   <chr>   <chr>            <chr>            <chr>            <chr>           
#> 1 L1      C294             Yuamen Xinneng … NA               NA              
#> 2 L2      C293             Yuamen Changyua… NA               NA              
#> 3 L3      C292             Yuama Ethanol L… IP5              Yuama Inc.      
#> 4 L4      C299             Yudaksel Holdin… NA               NA              
#> 5 L5      C305             Yukon Energy Co… NA               NA              
#> 6 L6      C304             Yukon Developme… NA               NA              
#> # … with 14 more variables: id_ultimate_parent <chr>,
#> #   name_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>