Dataset to bridge (translate) common sector-classification codes

This dataset serves as a translation key between common sector-classification systems and sectors relevant to the 'PACTA' tool (https://2degrees-investing.org/resource/pacta/).

naics_classification

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1057 rows and 4 columns.

Definitions

borderline (logical): Flag indicating if 2dii sector and classification code are a borderline match. The value TRUE indicates that the match is uncertain between the 2dii sector and the classification. The value FALSE indicates that the match is certainly perfect or the classification is certainly out of 2dii's scope..

code (double): Six-digit NAICS code.

naics_title (character): Original NAICS sector title.

sector (character): Associated 2dii sector.

Details

Classification datasets help to standardize sector classification codes from the wild to a relevant subset including 'power', 'oil and gas', 'coal', 'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.

See also

Examples

head(naics_classification)

#> # A tibble: 6 x 4
#>     code naics_title                      sector       borderline
#>    <dbl> <chr>                            <chr>        <lgl>     
#> 1 111110 soybean farming                  not in scope FALSE     
#> 2 111120 oilseed (except soybean) farming not in scope FALSE     
#> 3 111130 dry pea and bean farming         not in scope FALSE     
#> 4 111140 wheat farming                    not in scope FALSE     
#> 5 111150 corn farming                     not in scope FALSE     
#> 6 111160 rice farming                     not in scope FALSE