A view of available sector classification datasets

This dataset lists all sector classification code standards used by 'PACTA' (https://2degrees-investing.org/resource/pacta/).

sector_classifications

Format

An object of class tbl_df (inherits from tbl, data.frame) with 4700 rows and 4 columns.

Definitions

borderline (character): Flag indicating if 2dii sector and classification code are a borderline match. The value TRUE indicates that the match is uncertain between the 2dii sector and the classification. The value FALSE indicates that the match is certainly perfect or the classification is certainly out of 2dii's scope..

code (character): Formatted code.

code_system (character): Code system.

sector (character): Associated 2dii sector.

Details

Classification datasets help to standardize sector classification codes from the wild to a relevant subset including 'power', 'oil and gas', 'coal', 'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.

See also

Examples

head(sector_classifications)
#> # A tibble: 6 × 4
#>   sector       borderline code  code_system
#>   <chr>        <lgl>      <chr> <chr>      
#> 1 not in scope FALSE      100   CNB        
#> 2 not in scope FALSE      200   CNB        
#> 3 not in scope FALSE      300   CNB        
#> 4 not in scope FALSE      400   CNB        
#> 5 not in scope FALSE      500   CNB        
#> 6 not in scope FALSE      600   CNB