A loanbook dataset for demonstration

Fake financial portfolio.

Demo datasets are synthetic because most financial data is strictly private; they help to demonstrate and test the implementation in R of 'PACTA' (https://2degrees-investing.org/resource/pacta/).

loanbook_demo

Format

An object of class tbl_df (inherits from tbl, data.frame) with 321 rows and 19 columns.

Definitions

fi_type (character): Financial instrument name or asset class.

flag_project_finance_loan (character): Project finance flag denoting whether a loan is given out to a particular asset or not.

id_direct_loantaker (character): Borrower identifier unique to each borrower/sector combination in loanbook.

id_intermediate_parent_n (character): Optional input: id of the n-th intermediate parent company within the company structure that can be used for more granular mapping than the ultimate parent. Smaller values of n are closer to the direct_loantaker.

id_loan (character): Unique loan identifier.

id_ultimate_parent (character): Ultimate parent identifier unique to each ultimate parent/sector combination.

isin_direct_loantaker (logical): Optional input: providing the isin identifier of the direct loan taker to improve the matching coverage.

lei_direct_loantaker (logical): Optional input: providing the lei (legal entity identifier) of the direct loan taker to improve the matching coverage.

loan_size_credit_limit (double): Total credit limit or exposure at default.

loan_size_credit_limit_currency (character): Currency corresponding to credit limit.

loan_size_outstanding (double): Amount drawn by borrower from total credit limit.

loan_size_outstanding_currency (character): Currency corresponding to outstandings.

name_direct_loantaker (character): Name of the company directly taking the loan.

name_intermediate_parent_n (character): Optional input: name of intermediate parent company within the company structure that can be used for more granular mapping than the ultimate parent. Smaller values of n are closer to the direct_loantaker.

name_project (logical): Required input for loans with the flag_project_finance_loan = TRUE: Name of the project being financed.

name_ultimate_parent (character): Name of the ultimate parent company to which the borrower belongs. Can be the same as borrower.

sector_classification_direct_loantaker (double): Sector classification code of the direct loantaker.

sector_classification_input_type (character): Flag identifying if the sector classification code or character description is used.

sector_classification_system (character): Name of the sector classification standard being used.

See also

Examples

head(loanbook_demo)

#> # A tibble: 6 x 19
#>   id_loan id_direct_loanta… name_direct_loant… id_intermediate… name_intermedia…
#>   <chr>   <chr>             <chr>              <chr>            <chr>           
#> 1 L1      C294              Yuamen Xinneng Th… NA               NA              
#> 2 L2      C293              Yuamen Changyuan … NA               NA              
#> 3 L3      C292              Yuama Ethanol Llc  IP5              Yuama Inc.      
#> 4 L4      C299              Yudaksel Holding … NA               NA              
#> 5 L5      C305              Yukon Energy Corp… NA               NA              
#> 6 L6      C304              Yukon Development… NA               NA              
#> # … with 14 more variables: id_ultimate_parent <chr>,
#> #   name_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>